PROJ-151: Fix type bugs and cleanup across game, player, and packet files
- Change vel_x/vel_y from int32_t/uint32_t to float in Player, ClientRegularPacket, and AuthPlayerRecord
- Remove unsafe casts in update_player and form_auth_packet
- Remove duplicate game->status assignment in create_game
- Log error on partial map position read in create_game
- Fix typo "nauthorized" in handle_init_packet log message

+++ server/game/include/packet.h
@@ -63,8 +63,8 @@ struct __attribute__((packed)) ClientRegularPacket
     struct Header header;
     float         pos_x;
     float         pos_y;
-    int32_t       vel_x;
-    int32_t       vel_y;
+    float         vel_x;
+    float         vel_y;
     uint32_t      score;
 };
 
@@ -75,8 +75,8 @@ struct __attribute__((packed)) AuthPlayerRecord
     uint8_t  player_id[PLAYER_ID_SIZE];
     float    pos_x;
     float    pos_y;
-    int32_t  vel_x;
-    int32_t  vel_y;
+    float    vel_x;
+    float    vel_y;
     uint32_t score;
 };
